Exception handling

+++ src/main/java/ruijosecj/services/ProductService.java
@@ -3,14 +3,18 @@ package ruijosecj.services;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 import ruijosecj.dto.ProductDTO;
 import ruijosecj.entities.Product;
 import ruijosecj.repositories.ProductRepository;
+import ruijosecj.services.exceptions.DatabaseException;
+import ruijosecj.services.exceptions.ResourceNotFoundException;
 
 @Service
 public class ProductService {
@@ -21,7 +25,7 @@ public class ProductService {
 	@Transactional(readOnly = true)
 	public ProductDTO findById(Long id) {
 		Optional<Product> result = repository.findById(id);
-		Product product = result.get();
+		Product product = result.orElseThrow(()-> new ResourceNotFoundException("Recurso não encontrado!"));
 		ProductDTO dto = new ProductDTO(product);
 		return dto;
 	}
@@ -48,9 +52,17 @@ public class ProductService {
 		return new ProductDTO(entity);
 	}
 	
-	@Transactional
+	@Transactional(propagation = Propagation.SUPPORTS)
 	public void delete(Long id) {
-		repository.deleteById(id);
+		if (!repository.existsById(id)) {
+			throw new ResourceNotFoundException("Recurso não encontrado");
+		}
+		try {
+	        	repository.deleteById(id);    		
+		}
+	    	catch (DataIntegrityViolationException e) {
+	        	throw new DatabaseException("Falha de integridade referencial");
+	   	}
 	}
 
 	private void copyDtoToEntity(ProductDTO dto, Product entity) {
